modified:   lib/connect.py 	modified:   src/main.py

+++ src/main.py
@@ -13,7 +13,7 @@ time.sleep(3)
 import lib.connect as connect
 w=connect.Wifi()
 
-if w.run_mode():
+if w.creds() and w.set_sta(1) and w.run_mode():
     print("Normal mode activated - for chance to OS-mode type in terminal:")
     print(">>>import os")
     print(">>>os.remove('run_mode.dat'")    
@@ -21,4 +21,5 @@ if w.run_mode():
 else:
     print("OS mode activated")
     import web_os_run
+    web_os_run.run(w)
